matan feedback, add more tests.
