+++ dev/integration_tests/android_engine_test/lib/hcpp/platform_view_main.dart
@@ -10,6 +10,7 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_driver/driver_extension.dart';
 
+import '../platform_view/_shared.dart';
 import '../src/allow_list_devices.dart';
 
 void main() async {
@@ -18,59 +19,7 @@ void main() async {
 
   // Run on full screen.
   await SystemChrome.setEnabledSystemUIMode(SystemUiMode.immersive);
-  runApp(const MainApp());
-}
-
-// This should appear as the yellow line over a blue box. The
-// red box should not be visible unless the platform view has not loaded yet.
-final class MainApp extends StatefulWidget {
-  const MainApp({super.key});
-
-  @override
-  State<MainApp> createState() => _MainAppState();
-}
-
-class _MainAppState extends State<MainApp> {
-  bool showPlatformView = true;
-
-  void _togglePlatformView() {
-    setState(() {
-      showPlatformView = !showPlatformView;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      home: Stack(
-        alignment: AlignmentDirectional.center,
-        children: <Widget>[
-          TextButton(
-            key: const ValueKey<String>('AddOverlay'),
-            onPressed: _togglePlatformView,
-            child: const SizedBox(width: 190, height: 190, child: ColoredBox(color: Colors.green)),
-          ),
-          if (showPlatformView) ...<Widget>[
-            const SizedBox(
-              width: 200,
-              height: 200,
-              child: _HybridCompositionAndroidPlatformView(viewType: 'box_platform_view'),
-            ),
-            TextButton(
-              key: const ValueKey<String>('RemoveOverlay'),
-              onPressed: _togglePlatformView,
-              child: const SizedBox(
-                width: 800,
-                height: 25,
-                child: ColoredBox(color: Colors.yellow),
-              ),
-            ),
-          ],
-        ],
-      ),
-    );
-  }
+  runApp(const MainApp(platformView: _HybridCompositionAndroidPlatformView(viewType: 'box_platform_view')));
 }
 
 final class _HybridCompositionAndroidPlatformView extends StatelessWidget {
