+++ dev/integration_tests/android_engine_test/lib/platform_view/_shared.dart
@@ -4,18 +4,54 @@
 
 import 'package:flutter/material.dart';
 
-final class MainApp extends StatelessWidget {
+// This should appear as the yellow line over a blue box. The
+// green box should not be visible unless the platform view has not loaded yet.
+final class MainApp extends StatefulWidget {
   const MainApp({super.key, required this.platformView});
+
   final Widget platformView;
+
+  @override
+  State<MainApp> createState() => _MainAppState();
+}
+
+class _MainAppState extends State<MainApp> {
+  bool showPlatformView = true;
+
+  void _togglePlatformView() {
+    setState(() {
+      showPlatformView = !showPlatformView;
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       home: Stack(
+        alignment: AlignmentDirectional.center,
         children: <Widget>[
-          platformView,
-          Center(child: Container(width: 100, height: 100, color: Colors.red)),
+          TextButton(
+            key: const ValueKey<String>('AddOverlay'),
+            onPressed: _togglePlatformView,
+            child: const SizedBox(width: 190, height: 190, child: ColoredBox(color: Colors.green)),
+          ),
+          if (showPlatformView) ...<Widget>[
+            SizedBox(
+              width: 200,
+              height: 200,
+              child: widget.platformView,
+            ),
+            TextButton(
+              key: const ValueKey<String>('RemoveOverlay'),
+              onPressed: _togglePlatformView,
+              child: const SizedBox(
+                width: 800,
+                height: 25,
+                child: ColoredBox(color: Colors.yellow),
+              ),
+            ),
+          ],
         ],
       ),
     );
