+++ dev/integration_tests/android_engine_test/test_driver/hcpp/platform_view_main_test.dart
@@ -39,6 +39,8 @@ void main() async {
   });
 
   tearDownAll(() async {
+    await flutterDriver.tap(find.byValueKey('AddOverlay'));
+
     await nativeDriver.close();
     await flutterDriver.close();
   });
